update codes, upgrade BLINKER AT codes
